Refactor/ Restructure component layers for kebab button on containercard (#369)
* refactor: move DeleteContainerDialogContent up to the higher layer

* chore: delete unnecessary props

* chore: change the props name for readability

* chore: change props name for readability

* feat: implement useAutoFocus

* refactor: change the directory structure

* refactor: change the files and directory name

* refactor: change the file path

* docs: add TSDoc for interface
