+++ src/features/groups/components/ContainerCardList/ContainerCard/ContainerCardDropdownMenu/DeleteContainerDialogContent.tsx
